# contracts/openapi/arc.v1.yaml
openapi: 3.1.0
info:
  title: ARC API
  version: 1.0.0
  summary: MVP API for hot/cold archival storage
  description: Authoritative archive resources are durable across service restarts unless a specific contract explicitly expires or releases them.
servers:
  - url: /v1
paths:
  /collection-uploads:
    post:
      summary: Create or resume one collection upload session
      description: Creates or resumes the explicit resumable upload session for one collection id and expected logical file set.
      operationId: createOrResumeCollectionUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUploadRequest"
      responses:
        "200":
          description: Collection upload session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionUploadSession"
        default:
          $ref: "#/components/responses/Error"
  /collection-uploads/{collection_id}:
    get:
      summary: Get one collection upload session
      description: Returns the current resumable upload state for one collection id while it remains in progress. The session remains visible through `uploading`, `archiving`, and `failed`; once the collection archive package has uploaded, verified, and promoted the collection, this upload session is deleted and later reads return `not_found`.
      operationId: getCollectionUpload
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Explicit canonical collection id. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
      responses:
        "200":
          description: Collection upload session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionUploadSession"
        default:
          $ref: "#/components/responses/Error"
  /collection-uploads/{collection_id}/files/{path}/upload:
    post:
      summary: Create or resume one collection-file upload resource
      description: Creates or resumes one Riverhog-managed tus-compatible upload resource for one logical collection file until the published expiry time discards incomplete state.
      operationId: createOrResumeCollectionFileUpload
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Explicit canonical collection id. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
        - in: path
          name: path
          required: true
          description: Relative logical file path within the collection.
          schema: { type: string }
      responses:
        "200":
          description: Collection-file upload session
          headers:
            Tus-Resumable:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
            Upload-Offset:
              schema: { type: integer, minimum: 0 }
            Upload-Length:
              schema: { type: integer, minimum: 0 }
            Upload-Expires:
              schema: { type: string, format: date-time }
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionFileUploadSession"
        default:
          $ref: "#/components/responses/Error"
    patch:
      summary: Append one collection-file upload chunk
      description: Streams bytes into the resumable upload resource. When the final required chunk completes the last outstanding file and the full-file hashes verify, Riverhog moves the upload into `archiving`; the collection is not finalized or admitted until its whole-collection Glacier archive package uploads and verifies.
      operationId: appendCollectionFileUploadChunk
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Explicit canonical collection id. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
        - in: path
          name: path
          required: true
          description: Relative logical file path within the collection.
          schema: { type: string }
        - in: header
          name: Tus-Resumable
          required: true
          schema: { type: string, enum: ["1.0.0"] }
        - in: header
          name: Upload-Offset
          required: true
          schema: { type: integer, minimum: 0 }
        - in: header
          name: Upload-Checksum
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/offset+octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "204":
          description: Chunk accepted
          headers:
            Tus-Resumable:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
            Upload-Offset:
              schema: { type: integer, minimum: 0 }
            Upload-Length:
              schema: { type: integer, minimum: 0 }
            Upload-Expires:
              schema: { type: string, format: date-time }
        default:
          $ref: "#/components/responses/Error"
    head:
      summary: Read one collection-file upload resource state
      description: Returns the current tus-style offset and metadata for one existing Riverhog-managed collection-file upload resource. Returns `not_found` once the upload resource no longer exists.
      operationId: headCollectionFileUpload
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Explicit canonical collection id. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
        - in: path
          name: path
          required: true
          description: Relative logical file path within the collection.
          schema: { type: string }
      responses:
        "204":
          description: Current collection-file upload resource state
          headers:
            Tus-Resumable:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
            Upload-Offset:
              schema: { type: integer, minimum: 0 }
            Upload-Length:
              schema: { type: integer, minimum: 0 }
            Upload-Expires:
              schema: { type: string, format: date-time }
            Location:
              schema: { type: string }
        default:
          $ref: "#/components/responses/Error"
    delete:
      summary: Cancel one collection-file upload resource
      description: Cancels one existing Riverhog-managed collection-file upload resource, deletes any incomplete server-side bytes for that file, and resets that file back to `pending`.
      operationId: deleteCollectionFileUpload
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Explicit canonical collection id. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
        - in: path
          name: path
          required: true
          description: Relative logical file path within the collection.
          schema: { type: string }
      responses:
        "204":
          description: Upload resource canceled
          headers:
            Tus-Resumable:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
        default:
          $ref: "#/components/responses/Error"
    options:
      summary: Describe collection-file upload resource capabilities
      description: Returns the tus-compatible capability headers Riverhog supports for collection-file upload resources.
      operationId: optionsCollectionFileUpload
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Explicit canonical collection id. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
        - in: path
          name: path
          required: true
          description: Relative logical file path within the collection.
          schema: { type: string }
      responses:
        "204":
          description: Upload resource capabilities
          headers:
            Tus-Resumable:
              schema: { type: string }
            Tus-Version:
              schema: { type: string }
            Tus-Extension:
              schema: { type: string }
            Tus-Checksum-Algorithm:
              schema: { type: string }
        default:
          $ref: "#/components/responses/Error"
...
  /collections:
    get:
      summary: List collection summaries
      description: Returns one page of collection summaries with optional substring and protection-state filtering.
      operationId: listCollections
      parameters:
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: per_page
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - in: query
          name: q
          required: false
          description: Case-insensitive substring filter over canonical collection ids.
          schema: { type: string }
        - in: query
          name: protection_state
          required: false
          description: Exact collection protection filter.
          schema:
            type: string
            enum: [under_protected, cloud_only, physical_only, fully_protected]
      responses:
        "200":
          description: Collection summary listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionListResponse"
        default:
          $ref: "#/components/responses/Error"
  /collections/{collection_id}:
    get:
      summary: Get one collection summary
      operationId: getCollection
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Canonical collection id. Slash-bearing ids occupy multiple path segments under `/collections/`.
          schema: { type: string }
      responses:
        "200":
          description: Collection summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionSummary"
        default:
          $ref: "#/components/responses/Error"
...
components:
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    FetchEntryUploadState:
      type: string
      enum: [pending, partial, byte_complete, uploaded]
    PhysicalProtectionState:
      type: string
      enum: [unprotected, partially_protected, protected]
    CollectionProtectionState:
      type: string
      enum: [under_protected, cloud_only, physical_only, fully_protected]
    CoverageState:
      type: string
      enum: [none, partial, full]
    GlacierArchive:
      type: object
      required:
        [state, object_path, stored_bytes, backend, storage_class, last_uploaded_at, last_verified_at, failure]
      properties:
        state:
          type: string
          enum: [pending, uploading, uploaded, retrying, failed]
        object_path:
          oneOf:
            - type: string
            - type: "null"
        stored_bytes:
          oneOf:
            - type: integer
            - type: "null"
        backend:
          oneOf:
            - type: string
            - type: "null"
        storage_class:
          oneOf:
            - type: string
            - type: "null"
        last_uploaded_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        last_verified_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        failure:
          oneOf:
            - type: string
            - type: "null"
    ArchiveManifestProofState:
      type: string
      enum: [pending, uploaded, failed]
    CollectionArchiveManifest:
      type: object
      required: [object_path, sha256, ots_object_path, ots_state]
      properties:
        object_path:
          oneOf:
            - type: string
            - type: "null"
        sha256:
          oneOf:
            - type: string
              pattern: "^[0-9a-f]{64}$"
            - type: "null"
        ots_object_path:
          oneOf:
            - type: string
            - type: "null"
        ots_state:
          $ref: "#/components/schemas/ArchiveManifestProofState"
...
    CollectionSummary:
      type: object
      required:
        [id, files, bytes, hot_bytes, archived_bytes, pending_bytes, glacier, archive_manifest, archive_format, compression, disc_coverage, protection_state, protected_bytes, image_coverage]
      properties:
        id:
          type: string
          description: Canonical collection id. May contain `/` and is prefix-free with every other collection id.
        files: { type: integer }
        bytes: { type: integer }
        hot_bytes: { type: integer }
        archived_bytes: { type: integer }
        pending_bytes: { type: integer }
        glacier:
          $ref: "#/components/schemas/GlacierArchive"
        archive_manifest:
          $ref: "#/components/schemas/CollectionArchiveManifest"
        archive_format:
          type: string
        compression:
          type: string
        disc_coverage:
          $ref: "#/components/schemas/CollectionDiscCoverage"
        protection_state:
          $ref: "#/components/schemas/CollectionProtectionState"
        protected_bytes:
          type: integer
        image_coverage:
          type: array
          items:
            $ref: "#/components/schemas/CollectionCoverageImage"
    CollectionDiscCoverage:
      type: object
      required: [state, covered_bytes, verified_physical_bytes]
      properties:
        state:
          $ref: "#/components/schemas/CoverageState"
        covered_bytes:
          type: integer
          minimum: 0
        verified_physical_bytes:
          type: integer
          minimum: 0
    CollectionListResponse:
      type: object
      required: [page, per_page, total, pages, collections]
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
        collections:
          type: array
          items:
            $ref: "#/components/schemas/CollectionSummary"
    CollectionUploadFile:
      type: object
      required: [path, bytes, sha256]
      properties:
        path:
          type: string
          description: Relative logical file path within the collection source tree.
        bytes: { type: integer }
        sha256:
          type: string
          description: Logical-file SHA-256 digest for final verification.
    CollectionUploadRequest:
      type: object
      required: [collection_id, files]
      properties:
        collection_id:
          type: string
          description: Explicit canonical collection id for this upload session.
        ingest_source:
          oneOf:
            - type: string
            - type: "null"
          description: Descriptive source hint from the client. Never part of collection identity.
        files:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CollectionUploadFile"
    CollectionUploadFileState:
      type: object
      required: [path, bytes, sha256, upload_state, uploaded_bytes, upload_state_expires_at]
      properties:
        path: { type: string }
        bytes: { type: integer }
        sha256: { type: string }
        upload_state:
          $ref: "#/components/schemas/FetchEntryUploadState"
        uploaded_bytes:
          type: integer
          description: Accepted upload bytes for this logical file's plaintext stream.
        upload_state_expires_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
    CollectionUploadSession:
      type: object
      required:
        [
          collection_id,
          ingest_source,
          state,
          files_total,
          files_pending,
          files_partial,
          files_uploaded,
          bytes_total,
          uploaded_bytes,
          missing_bytes,
          upload_state_expires_at,
          files,
          collection,
        ]
      properties:
        collection_id:
          type: string
          description: Explicit canonical collection id for this upload session.
        ingest_source:
          oneOf:
            - type: string
            - type: "null"
        state:
          type: string
          enum: [uploading, archiving, finalized, failed]
        files_total: { type: integer }
        files_pending: { type: integer }
        files_partial: { type: integer }
        files_uploaded: { type: integer }
        bytes_total: { type: integer }
        uploaded_bytes: { type: integer }
        missing_bytes: { type: integer }
        upload_state_expires_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        files:
          type: array
          items:
            $ref: "#/components/schemas/CollectionUploadFileState"
        collection:
          oneOf:
            - $ref: "#/components/schemas/CollectionSummary"
            - type: "null"
    CollectionFileUploadSession:
      type: object
      required: [path, protocol, upload_url, offset, length, checksum_algorithm, expires_at]
      properties:
        path: { type: string }
        protocol:
          type: string
          const: tus
        upload_url:
          type: string
          format: uri
        offset:
          type: integer
          description: Current accepted byte offset in the logical file's upload stream.
        length:
          type: integer
          description: Total logical-file byte length for the upload stream.
        checksum_algorithm:
          type: string
          const: sha256
        expires_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
...
    CollectionCoverageImage:
      type: object
      required:
        [
          id,
          filename,
          physical_protection_state,
          physical_copies_required,
          physical_copies_registered,
          physical_copies_verified,
          physical_copies_missing,
          covered_paths,
          copies,
        ]
      properties:
        id:
          type: string
          pattern: "^[0-9]{8}T[0-9]{6}Z$"
        filename:
          type: string
        physical_protection_state:
          $ref: "#/components/schemas/PhysicalProtectionState"
        physical_copies_required:
          type: integer
          minimum: 1
        physical_copies_registered:
          type: integer
          minimum: 0
        physical_copies_verified:
          type: integer
          minimum: 0
        physical_copies_missing:
          type: integer
          minimum: 0
        covered_paths:
          type: array
          items:
            type: string
          description: Relative collection paths within this collection that the finalized image represents.
        copies:
          type: array
          items:
            $ref: "#/components/schemas/CopySummary"
...
    CopySummary:
      type: object
      required:
        [id, volume_id, label_text, location, created_at, state, verification_state, history]
      properties:
        id:
          type: string
          description: Generated copy id, unique within one finalized image/`volume_id`.
        volume_id:
          type: string
          pattern: "^[0-9]{8}T[0-9]{6}Z$"
          description: Immutable media identifier for the finalized image this copy belongs to.
        label_text:
          type: string
          description: Exact operator-facing disc label text for this generated copy slot.
        location:
          type: string
          nullable: true
          description: Mutable physical location label. Not part of copy identity.
        created_at: { type: string, format: date-time }
        state:
          type: string
          enum: [needed, burning, verified, registered, lost, damaged, retired]
        verification_state:
          type: string
          enum: [pending, verified, failed]
        history:
          type: array
          items:
            $ref: "#/components/schemas/CopyHistory"
    CopyHistory:
      type: object
      required: [at, event, state, verification_state, location]
      properties:
        at: { type: string, format: date-time }
        event: { type: string }
        state:
          type: string
          enum: [needed, burning, verified, registered, lost, damaged, retired]
        verification_state:
          type: string
          enum: [pending, verified, failed]
        location:
          type: string
          nullable: true
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }