# src/arc_api/routers/collections.py
from __future__ import annotations

from enum import StrEnum
from typing import Annotated

from fastapi import APIRouter, Query, Request, Response

from arc_api.deps import ContainerDep
from arc_api.mappers import map_collection, map_collection_list_page
from arc_api.schemas.collections import (
    CollectionFileUploadSessionOut,
    CollectionSummaryOut,
    CollectionUploadSessionOut,
    CreateOrResumeCollectionUploadRequest,
    ListCollectionsResponse,
)
from arc_api.tus import (
    tus_delete_headers,
    tus_options_headers,
    tus_upload_headers,
    validate_tus_chunk_request,
)

router = APIRouter(tags=["collections"])


class CollectionProtectionFilter(StrEnum):
    UNDER_PROTECTED = "under_protected"
    CLOUD_ONLY = "cloud_only"
    PHYSICAL_ONLY = "physical_only"
    FULLY_PROTECTED = "fully_protected"


_LEGACY_PROTECTION_FILTERS = {
    "under_protected": "partially_protected",
    "cloud_only": "unprotected",
    "physical_only": "partially_protected",
    "fully_protected": "protected",
}


@router.get("/collections", response_model=ListCollectionsResponse)
def list_collections(
    container: ContainerDep,
    page: int = Query(1, ge=1),
    per_page: int = Query(25, ge=1, le=100),
    q: str | None = Query(None),
    protection_state: Annotated[CollectionProtectionFilter | None, Query()] = None,
) -> ListCollectionsResponse:
    service_protection_state = (
        _LEGACY_PROTECTION_FILTERS[protection_state.value]
        if protection_state is not None
        else None
    )
    summary = container.collections.list(
        page=page,
        per_page=per_page,
        q=q,
        protection_state=service_protection_state,
    )
    return ListCollectionsResponse.model_validate(map_collection_list_page(summary))


@router.post("/collection-uploads", response_model=CollectionUploadSessionOut)
def create_or_resume_collection_upload(
    request: CreateOrResumeCollectionUploadRequest,
    container: ContainerDep,
) -> CollectionUploadSessionOut:
    payload = container.collections.create_or_resume_upload(
        collection_id=request.collection_id,
        files=[item.model_dump() for item in request.files],
        ingest_source=request.ingest_source,
    )
    return CollectionUploadSessionOut.model_validate(payload)


@router.get("/collection-uploads/{collection_id:path}", response_model=CollectionUploadSessionOut)
def get_collection_upload(
    collection_id: str,
    container: ContainerDep,
) -> CollectionUploadSessionOut:
    payload = container.collections.get_upload(collection_id)
    return CollectionUploadSessionOut.model_validate(payload)


@router.post(
    "/collection-uploads/{collection_id:path}/files/{path:path}/upload",
    response_model=CollectionFileUploadSessionOut,
)
def create_or_resume_collection_file_upload(
    collection_id: str,
    path: str,
    request: Request,
    response: Response,
    container: ContainerDep,
) -> CollectionFileUploadSessionOut:
    payload = container.collections.create_or_resume_file_upload(collection_id, path)
    payload["upload_url"] = str(request.url)
    response.headers.update(tus_upload_headers(payload, request=request))
    return CollectionFileUploadSessionOut.model_validate(payload)


@router.patch("/collection-uploads/{collection_id:path}/files/{path:path}/upload", status_code=204)
async def append_collection_file_upload_chunk(
    collection_id: str,
    path: str,
    request: Request,
    container: ContainerDep,
) -> Response:
    offset, checksum = validate_tus_chunk_request(request)
    payload = container.collections.append_upload_chunk(
        collection_id,
        path,
        offset=offset,
        checksum=checksum,
        content=await request.body(),
    )
    headers = tus_upload_headers(payload, request=request)
    return Response(status_code=204, headers=headers)


@router.head("/collection-uploads/{collection_id:path}/files/{path:path}/upload", status_code=204)
def head_collection_file_upload(
    collection_id: str,
    path: str,
    request: Request,
    container: ContainerDep,
) -> Response:
    payload = container.collections.get_file_upload(collection_id, path)
    return Response(status_code=204, headers=tus_upload_headers(payload, request=request))


@router.delete("/collection-uploads/{collection_id:path}/files/{path:path}/upload", status_code=204)
def delete_collection_file_upload(
    collection_id: str,
    path: str,
    container: ContainerDep,
) -> Response:
    container.collections.cancel_file_upload(collection_id, path)
    return Response(status_code=204, headers=tus_delete_headers())


@router.options(
    "/collection-uploads/{collection_id:path}/files/{path:path}/upload",
    status_code=204,
)
def options_collection_file_upload() -> Response:
    return Response(status_code=204, headers=tus_options_headers())


@router.get("/collections/{collection_id:path}", response_model=CollectionSummaryOut)
def get_collection(
    collection_id: str,
    container: ContainerDep,
) -> CollectionSummaryOut:
    summary = container.collections.get(collection_id)
    return CollectionSummaryOut.model_validate(map_collection(summary))